Add struct attribute regex definitions
This allows us to specify a regex on a tuple struct and have only the optional submatches be options. That is, if it matches, any capture that will always be present will be a `&'a str` while any that may or may not be there will be a `Option<&'a str>`.

In the future, I also want to add enums with top-level regex unions, possibly along with further parsing options (like trying to parse to a number).