Implement merging identical paths
Using suffix optimizations, we merge states with the same outgoing transitions. Then, when relevant, we de-duplicate incoming transitions (for diamond-pattern graphs that get simplified to a simple line). While this is possible to optimize away when squishing epsilons for a PikeVM engine state machine, doing it at the WorkingNFA stage allows the validity more optimizations (like the one-pass engine) to be detected in several cases.