Initial run-based optimizations
When we have a branchless sequence of ranges (a "run") we can do optimization to reduce code branching by avoiding short circuiting and generating code that's easier to vectorize. In the near future, I also want to look into whether implementing SIMD instructions would provide additional benefits here, though experiments would be necessary to determine if it's worth it.